// src/api/generated/models/lesson.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aegis Backend API
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  /** @nullable */
  closed_at?: string | null;
  course_id: string;
  created_at: string;
  ends_at: string;
  id: string;
  starts_at: string;
  title: string;
}